Demo commit message for Pretested Integration:
This is a longer commit message about what this commit is about.

We use this commit to illustrate the Prested Integration Plugin commit messages, based on their strategies. There are two demo jobs, one for each strategy and a job that does two commits, one on two different branches to be picked up as ready branches.