Single demo commit message for Pretested Integration:
This is a longer commit message about what this commit is about.

We use this commit to illustrate the Prested Integration Plugin commit messages, based on their strategies. There are two demo jobs, one for each strategy and jobs that does commits like this on ready branches to trigger the integration jobs.