#111: Single commit for Pretested Integration squashed:
This is a longer commit message about what this commit is about.
Strategy: squashed
This commit should be left untouched and fast-forwarded.
We use this commit to illustrate the Prested Integration Plugin commit messages, based on their strategies. There are two demo jobs, one for each strategy and jobs that does commits like this on ready branches to trigger the integration jobs.